Move `Timestamp` to `Calendar` namespace
