+++ tests/Calendar/CalendarConverterTest.php
@@ -11,7 +11,6 @@ use CultuurNet\UDB3\Calendar;
 use CultuurNet\UDB3\CalendarType;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\Hour;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\Minute;
-use CultuurNet\UDB3\Timestamp;
 use DateTime;
 use DateTimeInterface;
 use PHPUnit\Framework\TestCase;
